requirements.txt
python-dotenv>=1.0.0
openai>=1.10.0
pyautogen>=0.2.0
asyncio>=3.4.3
langchain>=0.0.267
langchain-openai>=0.0.1
pymongo>=4.5.0 
langchain-community